+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project are documented here. The format follows
 
 ## Unreleased
 
+### Changed
+
+- Notifications no longer play a sound.
+
 ## 0.2.0 - 2026-07-29
 
 ### Added